funcionales las rutas de productos
